Fixed and improved everything that remained to... Now LiveCD build script should have well working subscripts and main script. Improved documentation and feedback too. Deleted livecd-test.sh (replaced by livecd-edit.sh). I'm going to test the veracity of my sentences now... :-)
git-svn-id: http://svn.gna.org/svn/etoile/trunk/Etoile@2235 7b3f36aa-e6db-49c1-8b4c-66eac1997e64

+++ LiveCD/Subscripts/livecd-cleanup.sh
@@ -4,7 +4,16 @@
 # --cleanup
 #
 
-apt-get clean
+if [ -f $LIVECD_DIR/edit/etc/profile.original ]; then
+	echo
+	echo "LiveCD filesystem is in --test mode, you must first source "
+	echo "livecd-exit.sh in the test shell to exit properly"
+	echo
+	return
+fi
+
+sudo cp /etc/resolv.conf $LIVECD_DIR/edit/etc/
+sudo chroot $LIVECD_DIR/edit
 
 # Linux/GNOME specific cleanup
 rm -f /home/$ETOILE_USER_NAME/.bash_history
@@ -32,9 +41,12 @@ rm -r /home/$ETOILE_USER_NAME/GNUstep/Library/Addresses
 rm -r /home/$ETOILE_USER_NAME/GNUstep/Library/Bookmark
 
 # Build cleanup
+if [ $REMOVE_BUILD_FILES = yes ]; the
+	rm -r /build
+	apt-get -y uninstall subversion
+fi
 
-rm -r /build
-apt-get -y uninstall subversion
+apt-get clean
 
 exit
